# openapi/bookstore-api.yaml
openapi: 3.0.3
info:
  title: Bookstore API
  description: |
    REST API for the e-commerce bookstore application.
    Supports guest browsing, authenticated shopping, cart management,
    checkout, order history, and order cancellation.
  version: 1.0.0
  contact:
    name: Bookstore API Support
servers:
  - url: http://localhost:8080/api
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token. Obtain from /auth/login or /auth/register.

  # -------------------------------------------------------------------------
  # Reusable Parameters
  # -------------------------------------------------------------------------
  parameters:
    PageParam:
      name: page
      in: query
      description: Zero-based page number
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      description: Number of items per page
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    SortParam:
      name: sort
      in: query
      description: "Sort field and direction, e.g. price,asc or title,desc"
      schema:
        type: string
        example: title,asc

  # -------------------------------------------------------------------------
  # Reusable Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    PublisherResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Penguin Random House
        description:
          type: string
          nullable: true
        website:
          type: string
          format: uri
          nullable: true

    # --- Category -----------------------------------------------------------
    CategoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Science Fiction
        slug:
          type: string
          example: science-fiction
        description:
          type: string
          nullable: true
        parentId:
          type: string
          format: uuid
          nullable: true

    # --- Book ---------------------------------------------------------------
    BookSummaryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: "The Pragmatic Programmer"
        author:
          type: string
          example: "David Thomas, Andrew Hunt"
        coverImageUrl:
          type: string
          format: uri
          nullable: true
        price:
          type: number
          format: double
          example: 49.99
        averageRating:
          type: number
          format: double
          example: 4.7
        reviewCount:
          type: integer
          example: 312
        stockQuantity:
          type: integer
          example: 45
        category:
          $ref: '#/components/schemas/CategoryResponse'
        publisher:
          $ref: '#/components/schemas/PublisherResponse'
    BookDetailResponse:
      allOf:
        - $ref: '#/components/schemas/BookSummaryResponse'
        - type: object
          properties:
            isbn:
              type: string
              example: "978-0-13-595705-9"
            description:
              type: string
            pageCount:
              type: integer
              example: 352
            language:
              type: string
              example: English
            publishedDate:
              type: string
              format: date
              example: "2019-09-23"
            relatedBooks:
              type: array
              items:
                $ref: '#/components/schemas/BookSummaryResponse'
              maxItems: 6
...
    PageMetadata:
      type: object
      properties:
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          format: int64
          example: 148
        totalPages:
          type: integer
          example: 8
    PagedBookResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/BookSummaryResponse'
        pagination:
          $ref: '#/components/schemas/PageMetadata'

    # --- Cart ---------------------------------------------------------------
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Validation failed
        path:
          type: string
          example: /api/books
...
paths:
...
  /categories/{slug}/books:
    get:
      tags: [Categories]
      summary: List books in a category
      operationId: listBooksByCategory
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            example: science-fiction
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paged list of books in the category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedBookResponse'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==========================================================================
  # PUBLISHERS
  # ==========================================================================
...
  /books:
    get:
      tags: [Books]
      summary: Search and filter books
      operationId: searchBooks
      parameters:
        - name: query
          in: query
          description: Full-text search on title, author, ISBN, description
          schema:
            type: string
            example: "clean code"
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: publisherId
          in: query
          schema:
            type: string
            format: uuid
        - name: minPrice
          in: query
          schema:
            type: number
            format: double
        - name: maxPrice
          in: query
          schema:
            type: number
            format: double
        - name: minRating
          in: query
          schema:
            type: number
            format: double
        - name: inStockOnly
          in: query
          schema:
            type: boolean
            default: false
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paged search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedBookResponse'
  /books/{id}:
    get:
      tags: [Books]
      summary: Get book details
      operationId: getBookById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Full book details including related books
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDetailResponse'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==========================================================================
  # CART  (authenticated users only)
  # ==========================================================================
...